Refactor iOS implementation
Split out SwipeDismissContent as a reusable composable,
which in turns vastly simplifies the iOS implementation
of TiviNavigableCircuitContent.
